--- apps/fixed-borrow/src/ui/MintUSDCButton.tsx
@@ -1,40 +1,0 @@
-import { FaucetABI, SparkGoerliAddresses } from "@hyperdrive/spark";
-import { parseUnits } from "ethers/lib/utils.js";
-import { ReactElement } from "react";
-import { usePrepareContractWrite, useContractWrite, useAccount } from "wagmi";
-
-export function MintUSDCButton(): ReactElement {
-  const { address: account } = useAccount();
-  const { config: mintConfig } = usePrepareContractWrite({
-    address: SparkGoerliAddresses.faucet as `0x${string}`,
-    abi: FaucetABI,
-    functionName: "mint",
-    args: [
-      SparkGoerliAddresses.USDC_token as `0x${string}`,
-      account as `0x${string}`,
-      parseUnits("500", 8), // doesn't matter, faucet always mints 500 USDC
-    ],
-  });
-  const { write: mintTokens } = useContractWrite(mintConfig);
-  return (
-    <button
-      className="daisy-btn-outline daisy-btn-success daisy-btn mt-4"
-      onClick={() => mintTokens?.()}
-    >
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        viewBox="0 0 24 24"
-        fill="currentColor"
-        className="mr-2 h-6 w-6"
-      >
-        <path d="M10.464 8.746c.227-.18.497-.311.786-.394v2.795a2.252 2.252 0 01-.786-.393c-.394-.313-.546-.681-.546-1.004 0-.323.152-.691.546-1.004zM12.75 15.662v-2.824c.347.085.664.228.921.421.427.32.579.686.579.991 0 .305-.152.671-.579.991a2.534 2.534 0 01-.921.42z" />
-        <path
-          fillRule="evenodd"
-          d="M12 2.25c-5.385 0-9.75 4.365-9.75 9.75s4.365 9.75 9.75 9.75 9.75-4.365 9.75-9.75S17.385 2.25 12 2.25zM12.75 6a.75.75 0 00-1.5 0v.816a3.836 3.836 0 00-1.72.756c-.712.566-1.112 1.35-1.112 2.178 0 .829.4 1.612 1.113 2.178.502.4 1.102.647 1.719.756v2.978a2.536 2.536 0 01-.921-.421l-.879-.66a.75.75 0 00-.9 1.2l.879.66c.533.4 1.169.645 1.821.75V18a.75.75 0 001.5 0v-.81a4.124 4.124 0 001.821-.749c.745-.559 1.179-1.344 1.179-2.191 0-.847-.434-1.632-1.179-2.191a4.122 4.122 0 00-1.821-.75V8.354c.29.082.559.213.786.393l.415.33a.75.75 0 00.933-1.175l-.415-.33a3.836 3.836 0 00-1.719-.755V6z"
-          clipRule="evenodd"
-        />
-      </svg>
-      Mint 500 USDC
-    </button>
-  );
-}
